requirements.txt
Flask==3.0.3
matplotlib==3.9.2
numpy==2.1.1
Pillow==10.4.0
python-dotenv==1.0.1
redis==5.0.8
Requests==2.32.3
staticmap==0.5.7
tweepy==4.14.0
